+++ Classes/Model/Place.php
@@ -4,11 +4,7 @@ declare(strict_types=1);
 
 namespace OliverKlee\Seminars\Model;
 
-use OliverKlee\Oelib\Exception\NotFoundException;
-use OliverKlee\Oelib\Mapper\CountryMapper;
-use OliverKlee\Oelib\Mapper\MapperRegistry;
 use OliverKlee\Oelib\Model\AbstractModel;
-use OliverKlee\Oelib\Model\Country;
 
 /**
  * This class represents a place.
@@ -42,26 +38,5 @@ class Place extends AbstractModel
     public function getCity(): string
     {
         return $this->getAsString('city');
-    }
-
-    public function getCountry(): ?Country
-    {
-        $countryCode = $this->getAsString('country');
-        if ($countryCode === '') {
-            return null;
-        }
-
-        try {
-            $country = MapperRegistry::get(CountryMapper::class)->findByIsoAlpha2Code($countryCode);
-        } catch (NotFoundException $exception) {
-            $country = null;
-        }
-
-        return $country;
-    }
-
-    public function hasCountry(): bool
-    {
-        return $this->getCountry() !== null;
     }
 }